requirements.txt
streamlit>=1.32.0
langchain>=0.3.0
langchain-google-genai>=2.0.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
langsmith>=0.1.0
sqlite-vec>=0.1.1
pytesseract>=0.3.10
pdfplumber>=0.11.0
python-dotenv>=1.0.0
duckduckgo-search>=6.1.0
sentence-transformers>=3.0.0
Pillow>=10.0.0
langchain-text-splitters>=0.3.0
